+++ frontend/src/api/utils.js
@@ -56,6 +56,7 @@ export async function fetchJSON(url, opts) {
 
 export function adjustedData(data, url) {
   data.url = url;
+  data.hasUpdate = false;
   if (data.type === "directory") {
     if (!data.url.endsWith("/")) data.url += "/";