+++ frontend/src/notify/events.js
@@ -1,4 +1,5 @@
 import { mutations,state } from "@/store";
+import { notify } from "@/notify";
 
 export async function startSSE() {
 
@@ -10,15 +11,17 @@ export async function startSSE() {
 
     eventSrc.onerror = (err) => {
         mutations.updateSourceInfo("error");
-        mutations.setRealtimeActive(false);
-        console.log("SSE error:", err);
+        if (state.realtimeActive === true) {
+            mutations.setRealtimeActive(false);
+            notify.showError("The connection to server was lost. Trying to reconnect...");
+        }
     };
 
     eventSrc.onmessage = (event) => {
         try {
             const msg = JSON.parse(event.data);
             eventRouter(msg.eventType,msg.message);
-            console.log("Received event:", msg);
+            //console.log("Received event:", msg);
         } catch (err) {
             console.log("Error parsing event data:", err,event.data);
         }
@@ -28,10 +31,23 @@ export async function startSSE() {
 
 async function eventRouter(eventType,message) {
     switch (eventType) {
+        case "notification":
+            if (message == 'the server is shutting down') {
+                notify.showError("The server was shutdown - contact your admin if this was unexpected. Trying to reconnect...");
+                mutations.setRealtimeActive(false);
+                return
+            }
+            break
+        case "watchDirChange":
+            mutations.setWatchDirChangeAvailable(message);
+            break
         case "sourceUpdate":
             mutations.updateSourceInfo(message);
             break
         case "acknowledge":
+            if (state.realtimeActive === false) {
+                notify.showSuccess("The connection to server was re-established.");
+            }
             mutations.setRealtimeActive(true);
             break
         default: