+++ frontend/src/store/getters.js
@@ -3,6 +3,7 @@ import { getFileExtension } from  "@/utils/files.js";
 import { state } from "./state.js";
 import { noAuth } from "@/utils/constants.js";
 import { getTypeInfo } from "@/utils/mimetype";
+import * as i18n from "@/i18n";
 
 export const getters = {
   isScrollable: () => getters.currentView() === "settings" || getters.currentView() === "listingView",
@@ -20,6 +21,17 @@ export const getters = {
     return state.user.darkMode === true;
   },
   isLoggedIn: () => {
+    if (state.locale == "" || state.locale == undefined) {
+      const savedLocale = localStorage.getItem("userLocale");
+      if (savedLocale) {
+        i18n.setLocale(savedLocale);
+        i18n.default.locale = savedLocale;
+      } else {
+        const browserLocale = i18n.detectLocale();
+        i18n.setLocale(browserLocale);
+        i18n.default.locale = browserLocale;
+      }
+    }
     if (noAuth) {
       return true
     }